+++ common/src/main/java/com/alien/common/gameplay/item/PlatedNetherChitinArmorItem.java
@@ -1,33 +1,31 @@
-package com.human.common.gameplay.block_item;
+package com.alien.common.gameplay.item;
 
+import com.alien.common.registry.init.AlienArmorMaterials;
+import com.alien.common.registry.init.item.AlienArmorItems;
 import com.lib.common.data.TooltipHintBuilder;
 import com.lib.common.model.TooltipCategoryType;
-import net.minecraft.core.component.DataComponents;
 import net.minecraft.network.chat.Component;
-import net.minecraft.world.item.BlockItem;
-import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ArmorItem;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.TooltipFlag;
-import net.minecraft.world.item.component.ItemContainerContents;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
 
 import com.avp.common.data.TooltipTranslationKeys;
-import com.avp.common.registry.init.block.AVPBlocks;
 
-public class LeadChestBlockItem extends BlockItem {
+public class PlatedNetherChitinArmorItem extends ArmorItem {
 
     private static final List<Component> TOOLTIP_COMPONENTS = new TooltipHintBuilder()
-        .addCategory(TooltipCategoryType.WHEN_IN_INVENTORY)
-        .addPositiveEffect(TooltipTranslationKeys.EFFECT_AUTO_STORE_IRRADIATED_ITEMS)
+        .addCategory(TooltipCategoryType.WHEN_FULL_ARMOR_SET_EQUIPPED)
+        .addPositiveEffect(TooltipTranslationKeys.EFFECT_FIRE_RESISTANCE)
         .build();
 
-    public LeadChestBlockItem() {
+    public PlatedNetherChitinArmorItem(Type type) {
         super(
-            AVPBlocks.LEAD_CHEST.get(),
-            new Item.Properties().component(DataComponents.CONTAINER, ItemContainerContents.EMPTY)
-                .component(DataComponents.MAX_STACK_SIZE, 1)
+            AlienArmorMaterials.PLATED_NETHER_CHITIN.getHolder(),
+            type,
+            new Properties().durability(type.getDurability(AlienArmorItems.PLATED_CHITIN_DURABILITY_MULTIPLIER)).fireResistant()
         );
     }