+++ common/src/main/java/com/human/common/gameplay/block_item/AmmoChestBlockItem.java
@@ -1,6 +1,7 @@
 package com.human.common.gameplay.block_item;
 
-import net.minecraft.ChatFormatting;
+import com.lib.common.data.TooltipHintBuilder;
+import com.lib.common.model.TooltipCategoryType;
 import net.minecraft.core.component.DataComponents;
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.item.BlockItem;
@@ -12,9 +13,17 @@ import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
 
+import com.avp.common.data.TooltipTranslationKeys;
 import com.avp.common.registry.init.block.AVPBlocks;
 
 public class AmmoChestBlockItem extends BlockItem {
+
+    private static final List<Component> TOOLTIP_COMPONENTS = new TooltipHintBuilder()
+        .addCategory(TooltipCategoryType.WHEN_IN_INVENTORY)
+        .addPositiveEffect(TooltipTranslationKeys.EFFECT_GUNS_AUTO_RELOAD_FROM_CHEST)
+        .addCategory(TooltipCategoryType.WHEN_PLACED_IN_WORLD)
+        .addPositiveEffect(TooltipTranslationKeys.EFFECT_NEARBY_TURRETS_USE_AMMO_FROM_CHEST)
+        .build();
 
     public AmmoChestBlockItem() {
         super(
@@ -32,22 +41,6 @@ public class AmmoChestBlockItem extends BlockItem {
         @NotNull TooltipFlag tooltipFlag
     ) {
         super.appendHoverText(stack, context, tooltipComponents, tooltipFlag);
-
-        tooltipComponents.add(
-            Component.translatable("tooltip.avp.ammo_chest.in_inventory")
-                .withStyle(ChatFormatting.YELLOW)
-        );
-        tooltipComponents.add(
-            Component.translatable("tooltip.avp.ammo_chest.reload_from_chest")
-                .withStyle(ChatFormatting.GREEN)
-        );
-        tooltipComponents.add(
-            Component.translatable("tooltip.avp.ammo_chest.placed")
-                .withStyle(ChatFormatting.YELLOW)
-        );
-        tooltipComponents.add(
-            Component.translatable("tooltip.avp.ammo_chest.turret_load_from_chest")
-                .withStyle(ChatFormatting.GREEN)
-        );
+        tooltipComponents.addAll(TOOLTIP_COMPONENTS);
     }
 }