+++ common/src/main/java/com/human/common/gameplay/item/MK50ArmorItem.java
@@ -1,6 +1,7 @@
 package com.human.common.gameplay.item;
 
-import net.minecraft.ChatFormatting;
+import com.lib.common.data.TooltipHintBuilder;
+import com.lib.common.model.TooltipCategoryType;
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.item.ArmorItem;
 import net.minecraft.world.item.Item;
@@ -10,11 +11,19 @@ import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
 
+import com.avp.common.data.TooltipTranslationKeys;
 import com.avp.common.registry.init.AVPArmorMaterials;
 
 public class MK50ArmorItem extends ArmorItem {
 
     private static final int MK50_DURABILITY_MULTIPLIER = 14;
+
+    private static final List<Component> TOOLTIP_COMPONENTS = new TooltipHintBuilder()
+        .addCategory(TooltipCategoryType.WHEN_FULL_ARMOR_SET_EQUIPPED)
+        .addPositiveEffect(TooltipTranslationKeys.EFFECT_RADIATION_RESISTANCE)
+        .addPositiveEffect(TooltipTranslationKeys.EFFECT_WATER_BREATHING)
+        .addNegativeEffect(TooltipTranslationKeys.EFFECT_SLOWNESS)
+        .build();
 
     public MK50ArmorItem(ArmorItem.Type type) {
         super(AVPArmorMaterials.MK50.getHolder(), type, new Item.Properties().durability(type.getDurability(MK50_DURABILITY_MULTIPLIER)));
@@ -28,22 +37,6 @@ public class MK50ArmorItem extends ArmorItem {
         @NotNull TooltipFlag tooltipFlag
     ) {
         super.appendHoverText(stack, context, tooltipComponents, tooltipFlag);
-
-        tooltipComponents.add(
-            Component.translatable("tooltip.avp.mk50_suit.full_set_bonus")
-                .withStyle(ChatFormatting.YELLOW)
-        );
-        tooltipComponents.add(
-            Component.translatable("tooltip.avp.mk50_suit.radiation_resistance")
-                .withStyle(ChatFormatting.GREEN)
-        );
-        tooltipComponents.add(
-            Component.translatable("tooltip.avp.mk50_suit.water_breathing")
-                .withStyle(ChatFormatting.GREEN)
-        );
-        tooltipComponents.add(
-            Component.translatable("tooltip.avp.mk50_suit.slowness")
-                .withStyle(ChatFormatting.RED)
-        );
+        tooltipComponents.addAll(TOOLTIP_COMPONENTS);
     }
 }