+++ common/src/main/java/com/predator/common/gameplay/item/VeritaniumArmorItem.java
@@ -1,33 +1,35 @@
-package com.human.common.gameplay.block_item;
+package com.predator.common.gameplay.item;
 
 import com.lib.common.data.TooltipHintBuilder;
 import com.lib.common.model.TooltipCategoryType;
-import net.minecraft.core.component.DataComponents;
+import com.predator.common.registry.init.PredatorArmorMaterials;
 import net.minecraft.network.chat.Component;
-import net.minecraft.world.item.BlockItem;
-import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ArmorItem;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.TooltipFlag;
-import net.minecraft.world.item.component.ItemContainerContents;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
 
 import com.avp.common.data.TooltipTranslationKeys;
-import com.avp.common.registry.init.block.AVPBlocks;
 
-public class LeadChestBlockItem extends BlockItem {
+public class VeritaniumArmorItem extends ArmorItem {
+
+    private static final int VERITANIUM_DURABILITY_MULTIPLIER = 40;
 
     private static final List<Component> TOOLTIP_COMPONENTS = new TooltipHintBuilder()
-        .addCategory(TooltipCategoryType.WHEN_IN_INVENTORY)
-        .addPositiveEffect(TooltipTranslationKeys.EFFECT_AUTO_STORE_IRRADIATED_ITEMS)
+        .addCategory(TooltipCategoryType.WHEN_HELMET_EQUIPPED)
+        .addPositiveEffect(TooltipTranslationKeys.EFFECT_PREVENTS_FACEHUGGING)
+        .addCategory(TooltipCategoryType.WHEN_FULL_ARMOR_SET_EQUIPPED)
+        .addPositiveEffect(TooltipTranslationKeys.EFFECT_JUMP_BOOST)
+        .addPositiveEffect(TooltipTranslationKeys.EFFECT_STRENGTH)
         .build();
 
-    public LeadChestBlockItem() {
+    public VeritaniumArmorItem(Type type) {
         super(
-            AVPBlocks.LEAD_CHEST.get(),
-            new Item.Properties().component(DataComponents.CONTAINER, ItemContainerContents.EMPTY)
-                .component(DataComponents.MAX_STACK_SIZE, 1)
+            PredatorArmorMaterials.VERITANIUM.getHolder(),
+            type,
+            new Properties().durability(type.getDurability(VERITANIUM_DURABILITY_MULTIPLIER)).fireResistant()
         );
     }